DPAD-1411 :: Fixed compilation issues with ts-node due to a higher than intended zx version | Added some additional logging
