+++ package.json
@@ -14,7 +14,7 @@
 		"build": "yarn build-fandom-player && yarn build-stand-alone",
 		"build-fandom-player": "yarn run create-dist && rollup -c && cp package.json dist/ && cd dist && npm set-script prepare ''",
 		"build-stand-alone": "yarn run create-standalone-dist && rollup -c rollup-stand-alone.config.js && cp package.json standalone-dist && npm set-script prepare ''",
-		"deploy-stand-alone": "ts-node --esm deploy/deploy.ts",
+		"deploy-stand-alone": "ts-node --project tsconfig-stand-alone.json src/deploy/deploy.ts",
 		"watch": "yarn run create-dist && NODE_ENV=development rollup -c --watch",
 		"app": "cd test-jw && yarn && yarn start",
 		"create-dist": "rm -rf dist && mkdir -p dist",
@@ -31,7 +31,7 @@
 		"lint:eslint:tests": "eslint --cache --fix tests",
 		"lint:eslint": "npm-run-all --continue-on-error --parallel lint:eslint:src",
 		"test": "echo 'No Tests yet'",
-		"prepare": "husky install",
+		"prepare": "",
 		"pub:patch": "npm version patch && rm -rf dist && yarn pub",
 		"pub:minor": "npm version minor && rm -rf dist && yarn pub",
 		"pub:major": "npm version major && rm -rf dist && yarn pub",
@@ -108,12 +108,13 @@
 		"rollup-plugin-terser": "7.0.2",
 		"tslib": "^2.3.1",
 		"ts-node": "^10.4.0",
+		"tsconfig-paths": "^3.12.0",
 		"typescript": "^4.5.4",
 		"i18next": "^19.7.0",
 		"react": "^16.14.0 || ^17.0.2 || ^18.0.0",
 		"react-dom": "^16.14.0 || ^17.0.2 || ^18.0.0",
 		"react-i18next": "^11.7.1",
-		"zx": "^7.1.1",
+		"zx": "^4.2.0",
 		"@google-cloud/storage": "^5.16.1"
 	},
 	"peerDependencies": {
