+++ src/deploy/deployToEnv.ts
@@ -1,6 +1,6 @@
 import { $ } from 'zx';
 
-import {Env, DESTINATION_MAPPING, VIDEO_PREFIX} from './envs';
+import { Env, DESTINATION_MAPPING, VIDEO_PREFIX, VIDEO_DIR } from './envs';
 import { uploadDistTo } from './gcs';
 
 function getGCSDestination(location: string): string {
@@ -9,17 +9,15 @@ function getGCSDestination(location: string): string {
 		process.exit(1);
 	}
 
-
-	return String(`${VIDEO_PREFIX}/${DESTINATION_MAPPING[location]}`);
+	return String(`${VIDEO_PREFIX}/${DESTINATION_MAPPING[location]}/${VIDEO_DIR}`);
 }
 
-export async function deployToEnv(
-	env: Env,
-) {
+export async function deployToEnv(env: Env) {
 	await $`yarn build-stand-alone`;
 
 	// Uploading to Google Cloud Storage
 	const gCSLocation = getGCSDestination(env);
+	console.log(`gCSLocation set to - ${gCSLocation}`);
 	console.log(`Uploading GCS to ${env}`);
 	await uploadDistTo(gCSLocation, env);
 }
