+++ tsconfig-stand-alone.json
@@ -21,8 +21,14 @@
   "ts-node": {
     // these options are overrides used only by ts-node
     // same as the --compilerOptions flag and the TS_NODE_COMPILER_OPTIONS environment variable
+    "transpileOnly": true,
+    "files": true,
+    "require": ["tsconfig-paths/register"],
     "compilerOptions": {
-      "module": "es2020"
+      "module": "commonjs",
+      "target": "es2017",
+      "esModuleInterop": true,
+      "skipLibCheck": true
     }
   }
 }
